backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
scikit-learn==1.5.2
xgboost>=2.0.0
lightgbm>=4.3.0
shap==0.45.0
imbalanced-learn>=0.12.0
pandas>=2.2.0
numpy==1.26.4
dowhy==0.11.1
scikit-survival==0.23.0
transformers>=4.40.0
torch>=2.2.0
SQLAlchemy>=2.0.28
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
neurokit2==0.2.6
catboost>=1.2.0
easyocr>=1.7.0
pdfplumber>=0.11.0
opencv-python>=4.9.0
google-generativeai>=0.4.1
fastai>=2.7.14
joblib>=1.3.2
Pillow>=10.2.0
python-dotenv>=1.0.1
